+++ src/create.test.js
@@ -1,4 +1,5 @@
-module.exports = function (options) {
+/* global assert:true */
+export default function (options) {
   describe('Adapter#create', function () {
     it('should exist', function () {
       assert.equal(typeof this.$$adapter.create, 'function', 'adapter should have a "create" method')
@@ -8,20 +9,21 @@ module.exports = function (options) {
       const User = this.$$User
       const props = { name: 'John' }
 
-      assert.debug('create', props)
+      assert.debug('create', User.name, props)
       const user = await adapter.create(User, props)
-      assert.debug('created', JSON.stringify(user, null, 2))
+      const userId = user[User.idAttribute]
+      assert.debug('created', User.name, user)
 
-      assert.equal(user.name, props.name, `name of user should be "${props.name}"`)
-      assert.isDefined(user[User.idAttribute], 'new user should have an id')
+      assert.equal(user.name, props.name, 'user.name')
+      assert.isDefined(user[User.idAttribute], 'user[User.idAttribute]')
 
-      assert.debug('find', user[User.idAttribute])
-      const foundUser = await adapter.find(User, user[User.idAttribute])
-      assert.debug('found', JSON.stringify(foundUser, null, 2))
+      assert.debug('find', User.name, userId)
+      const foundUser = await adapter.find(User, userId)
+      assert.debug('found', User.name, foundUser)
 
-      assert.equal(foundUser.name, props.name, `name of user should be "${props.name}"`)
-      assert.isDefined(foundUser[User.idAttribute], 'new user should have an id')
-      assert.equal(foundUser[User.idAttribute], user[User.idAttribute])
+      assert.equal(foundUser.name, props.name, 'foundUser.name')
+      assert.isDefined(foundUser[User.idAttribute], 'foundUser[User.idAttribute]')
+      assert.equal(foundUser[User.idAttribute], userId, 'foundUser[User.idAttribute]')
     })
   })
 }