+++ src/updateAll.test.js
@@ -1,4 +1,5 @@
-module.exports = function (options) {
+/* global assert:true */
+export default function (options) {
   describe('Adapter#updateAll', function () {
     it('should exist', function () {
       assert.equal(typeof this.$$adapter.updateAll, 'function', 'adapter should have a "updateAll" method')
@@ -8,21 +9,21 @@ module.exports = function (options) {
       const User = this.$$User
       let props = { name: 'John', age: 20 }
 
-      assert.debug('create', props)
+      assert.debug('create', User.name, props)
       const user1 = await adapter.create(User, props)
-      assert.debug('created', JSON.stringify(user1, null, 2))
+      assert.debug('created', User.name, user1)
       const userId1 = user1[User.idAttribute]
 
       props = { name: 'John', age: 30 }
 
-      assert.debug('create', props)
+      assert.debug('create', User.name, props)
       const user2 = await adapter.create(User, props)
-      assert.debug('created', JSON.stringify(user2, null, 2))
+      assert.debug('created', User.name, user2)
       const userId2 = user2[User.idAttribute]
 
-      assert.debug('findAll', { name: 'John' })
+      assert.debug('findAll', User.name, { name: 'John' })
       const users = await adapter.findAll(User, { name: 'John' })
-      assert.debug('found', JSON.stringify(users, null, 2))
+      assert.debug('found', User.name, users)
       users.sort(function (a, b) {
         return a.age - b.age
       })
@@ -33,9 +34,9 @@ module.exports = function (options) {
       assert.equal(users.filter(function (x) { return x.age === 20 }).length, 1)
       assert.equal(users.filter(function (x) { return x.age === 30 }).length, 1)
 
-      assert.debug('updateAll', { name: 'Johnny' }, { name: 'John' })
+      assert.debug('updateAll', User.name, { name: 'Johnny' }, { name: 'John' })
       const users2 = await adapter.updateAll(User, { name: 'Johnny' }, { name: 'John' })
-      assert.debug('updated', JSON.stringify(users2, null, 2))
+      assert.debug('updated', User.name, users2)
       users2.sort(function (a, b) {
         return a.age - b.age
       })
@@ -46,15 +47,16 @@ module.exports = function (options) {
       assert.equal(users2.filter(function (x) { return x.age === 20 }).length, 1)
       assert.equal(users2.filter(function (x) { return x.age === 30 }).length, 1)
 
-      assert.debug('findAll', { name: 'John' })
+      assert.debug('findAll', User.name, { name: 'John' })
       const users3 = await adapter.findAll(User, { name: 'John' })
-      assert.debug('found', JSON.stringify(users3, null, 2))
+      assert.debug('found', User.name, users3)
       assert.equalObjects(users3, [])
       assert.equal(users3.length, 0)
 
-      assert.debug('findAll', { name: 'Johnny' })
+      assert.debug('findAll', User.name, { name: 'Johnny' })
       const users4 = await adapter.findAll(User, { name: 'Johnny' })
-      assert.debug('found', JSON.stringify(users4, null, 2))
+      assert.debug('found', User.name, users4)
+
       users4.sort(function (a, b) {
         return a.age - b.age
       })