+++ CHANGELOG.md
@@ -1,6 +1,90 @@
+##### 2.0.0-alpha.21 - 03 May 2016
+
+- Fix peer dependencies
+
+##### 2.0.0-alpha.20 - 02 May 2016
+
+- Fix for SQL testing
+
+##### 2.0.0-alpha.19 - 02 May 2016
+
+- Fix for SQL testing
+
+##### 2.0.0-alpha.18 - 02 May 2016
+
+- Fix for SQL testing
+
+##### 2.0.0-alpha.17 - 30 April 2016
+
+- Fix for testing in IE9
+
+##### 2.0.0-alpha.16 - 17 March 2016
+
+- Added tests for "count" and "sum" methods
+
+##### 2.0.0-alpha.15 - 12 March 2016
+
+- Added tests for filtering in "with" sub queries
+
+##### 2.0.0-alpha.14 - 08 March 2016
+
+- Features and Methods are now opt-out rather than opt-in
+
+##### 2.0.0-alpha.13 - 03 March 2016
+
+- Added tests for subclassing adapters
+
+##### 2.0.0-alpha.12 - 01 March 2016
+
+- Some improved testing
+
+##### 2.0.0-alpha.11 - 29 February 2016
+
+- Some improved testing
+
+##### 2.0.0-alpha.10 - 26 February 2016
+
+- Some fixes
+
+##### 2.0.0-alpha.9 - 26 February 2016
+
+- Some fixes
+
+##### 2.0.0-alpha.8 - 26 February 2016
+
+- destroy and destroyAll no longer have return values
+
+##### 2.0.0-alpha.7 - 26 February 2016
+
+- Improved testing
+
+##### 2.0.0-alpha.6 - 23 February 2016
+
+- Improved testing of find, destroy, and update features
+
+##### 2.0.0-alpha.5 - 18 February 2016
+
+- rebased 1.5.0 changes
+
 ##### 1.5.0 - 18 February 2016
 
 - Some improvements
+
+##### 2.0.0-alpha.4 - 16 February 2016
+
+- Better configurability
+
+##### 2.0.0-alpha.3 - 16 February 2016
+
+- Added source map
+
+##### 2.0.0-alpha.2 - 16 February 2016
+
+- Soem fixes
+
+##### 2.0.0-alpha.1 - 13 February 2016
+
+- Upgrades to work with 3.x js-data and adapters
 
 ##### 1.3.0 - 10 November 2015